% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{davies.pwr}
\alias{davies.pwr}
\title{Compute power for SKAT (given chi-square mixture parameters) using Davies method}
\usage{
davies.pwr(alpha, lambda, delta)
}
\arguments{
\item{alpha}{significance level}

\item{lambda}{vector of mixing coefficients}

\item{delta}{vector of non-centrality parameters}
}
\value{
estimated power
}
\description{
The Davies method is based on inverting characteristic function to compute the exact distribution function.
}
\references{
Wu,B. and Pankow,J.S. (2016) On sample size and power calculation for variant set-based association tests. /AHG/, 80(2), 136-143.

Lafaye De Micheaux, P. (2013) CompQuadForm: Distribution function of quadratic forms in normal variables. R package version 1.4.1.
}
