% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{lee.pwr}
\alias{lee.pwr}
\title{Compute power for SKAT (given chi-square mixture parameters) using non-central chi-square approximation}
\usage{
lee.pwr(alpha, lambda, delta)
}
\arguments{
\item{alpha}{significance level}

\item{lambda}{vector of mixing coefficients}

\item{delta}{vector of non-centrality parameters}
}
\value{
estimated power
}
\description{
The non-central chi-square approximation is based on a modified Liu method.
}
\references{
Wu,B. and Pankow,J.S. (2016) On sample size and power calculation for variant set-based association tests. /AHG/, 80(2), 136-143.

Wu,B. and Pankow,J.S. (2017) On computing the tail probability of non-negative definite quadratic forms in central normal variables. tech rep.

Lee, S., Wu, M. C., and Lin, X. (2012) Optimal tests for rare variant effects in sequencing association studies. Biostatistics, 13, 762-775.
}
