% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{liu.pwr}
\alias{liu.pwr}
\title{Compute power for SKAT (given chi-square mixture parameters) using non-central chi-square approximation}
\usage{
liu.pwr(alpha, lambda, delta)
}
\arguments{
\item{alpha}{significance level}

\item{lambda}{vector of mixing coefficients}

\item{delta}{vector of non-centrality parameters}
}
\value{
estimated power
}
\description{
The non-central chi-square approximation is based on the Liu method.
}
\references{
Wu,B. and Pankow,J.S. (2016) On sample size and power calculation for variant set-based association tests. /AHG/, 80(2), 136-143.

Wu,B. and Pankow,J.S. (2017) On computing the tail probability of non-negative definite quadratic forms in central normal variables. tech rep.

H. Liu, Y. Tang, H.H. Zhang (2009) A new chi-square approximation to the distribution of non-negative definite quadratic forms in non-central normal variables, Computational Statistics and Data Analysis, 53, 853-856.
}
