% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{RVS.params}
\alias{RVS.params}
\title{Compute parameters for (weighted linear) kernel association test of rare variant set}
\usage{
RVS.params(n = 2000, VLD, Ves, MAF, Wv = NULL, W.beta = c(1, 25))
}
\arguments{
\item{n}{sample size of a planned study}

\item{VLD}{population genotype covariance matrix for the variant set}

\item{Ves}{variant effect sizes}

\item{MAF}{minor allele freqs of variant set}

\item{Wv}{variant weights}

\item{W.beta}{beta dist parameters to compute variant weights based on MAF when Wv is not specified.}
}
\value{
\describe{
  \item{lambda}{ mixing cofficients of 1-DF chi-square mixtures }
  \item{delta}{ non-centrality parameters of 1-DF chi-square mixtures }
}
}
\description{
Assume a linear model for continous traits, \eqn{Y=\alpha+G\beta+\epsilon}.
Input population genotype covariance matrix (VLD), variant weights (Wv), variant effect sizes (Ves, i.e., \eqn{\beta}), and sample size (n).
This function computes the asymptotic mixture of 1-DF chi-square distributions, \eqn{\sum_i\lambda_i\chi_1^2(\delta_i)}.
}
\references{
Wu,B. and Pankow,J.S. (2016) On sample size and power calculation for variant set-based association tests. /AHG/, 80(2), 136-143.
}
