% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{wu.pwr}
\alias{wu.pwr}
\title{Compute power for SKAT (given chi-square mixture parameters) using non-central chi-square approximation}
\usage{
wu.pwr(alpha, lambda, delta, N0 = 12, N1 = 6)
}
\arguments{
\item{alpha}{significance level}

\item{lambda}{vector of mixing coefficients}

\item{delta}{vector of non-centrality parameters}

\item{N0}{match (N0-1,N0)-th moments for null p-value calculation}

\item{N1}{match (N1-1,N1)-th moments for power calculation}
}
\value{
\describe{
  \item{pwr}{ estimated power }
}
}
\description{
Given the 1-DF chi-square mixture distribution, \eqn{\sum_i\lambda_i\chi_1^2(\delta_i)},
the non-central chi-square approximation is based on moment-matching (typically higher moment). See the reference.
}
\references{
Wu,B. and Pankow,J.S. (2016) On sample size and power calculation for variant set-based association tests. /AHG/, 80(2), 136-143.

Wu,B. and Pankow,J.S. (2017) On computing the tail probability of non-negative definite quadratic forms in central normal variables. tech rep.
}
